ml-optimization/requirements.txt
# Core ML/Optimization Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning & RL
tensorflow>=2.13.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Gradient Boosting
xgboost>=1.7.0
lightgbm>=4.0.0

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
redis>=5.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0

# Monitoring & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0


